pyproject.toml
[tool.poetry]
name = "project-4"
version = "0.1.0"
description = ""
authors = ["hhssome <artempavel-terehove@mail.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy2-stubs = "^0.0.2a29"
sqlalchemy = "^1.4.44"
fastapi = "^0.88.0"
pydantic = "^1.10.2"
python-jose = "^3.3.0"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
shortuuid = "^1.0.11"
pytest = "^7.2.0"
uvicorn = "^0.20.0"
asyncpg = "^0.27.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
orjson = "^3.8.3"
httpx = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
